chore(clippy): scripted lint pass — collapsible_if/assign_op/doc_lazy + safe expr rewrites
Applied clippy's own machine/maybe-incorrect suggestions verbatim for safe
mechanical lints across in-scope crates (collapsible_if, assign_op_pattern,
manual_is_multiple_of, clone_on_copy, identity_op, doc_lazy_continuation,
doc_overindented, unused_parens, for_kv_map, unused_enumerate_index,
unnecessary_to_owned/map_or/sort_by, redundant_closure, needless_borrow, etc.).
Behavior unchanged; all in-scope crates build --all-targets green.

+++ observability/src/main.rs
@@ -255,7 +255,7 @@ fn emit_turn_lifecycle(em: &Emitter) -> [u8; 32] {
                     post_state_hash: hex32(&receipt.post_state_hash),
                     effects_hash: hex32(&receipt.effects_hash),
                     timestamp: receipt.timestamp,
-                    action_count: receipt.action_count as usize,
+                    action_count: receipt.action_count,
                     computrons_used,
                     finality: format!("{:?}", receipt.finality),
                 },
